Remove dimensions from bounding box selections to ease scripting
Bounding box dimensions are based on the image size, which likely changes from one image to the next. As scripting requires a match to the exact selection, scripts will not typically know the dimensions and will not find the selection. This update simplifies selection identification by removing these dimensions from the choices strings.